fix(ci): resolve persistent e2e-test failures in GitHub Actions
- Remove --frozen-lockfile (inconsistent with other jobs)
- Use build+preview instead of dev server in CI (faster, deterministic)
- Add retries=2 and increased timeouts for CI stability
- Fix e2e tests for new route structure (/orders→/trading, /backtest→/research)
- Upload Playwright report artifact for debugging failures
- Add e2e-test dependency on web-build job
- Fix TradeRecord quantity type from float to int

+++ src/api/routes/backtest.py
@@ -239,7 +239,7 @@ async def get_backtest_result(task_id: str, api_key: str = Depends(verify_api_ke
             date=str(t.timestamp),
             symbol=t.symbol,
             side=t.side.value,
-            quantity=float(t.quantity),
+            quantity=int(t.quantity),
             price=float(t.price),
             commission=float(t.commission),
         )
